Refactoring folders and import in alertDetails
