+++ src/components/Alerts/AlertDetails/AlertInfos/AlertInfosSection.tsx
