+++ src/components/Alerts/AlertDetails/AlertInfos/AlertMap.tsx
